requirements.txt
asgiref==3.8.1
boto3==1.35.98
botocore==1.35.98
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cryptography==44.0.1
decouple==0.0.7
Django==5.1.4
django-cors-headers==4.6.0
django-extensions==3.2.3
django-ratelimit==4.1.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
mercadopago==2.2.3
olefile==0.47
packaging==24.2
paypalrestsdk==1.13.3
pillow==11.0.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
reportlab==4.3.0
requests==2.32.3
s3transfer==0.10.4
secure-smtplib==0.1.1
six==1.17.0
sqlparse==0.5.3
stripe==11.5.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
